move encode/decoded functions to appropriate client-side vs server-side directories
