+++ utils/github/getDecodedData.ts
@@ -1,7 +1,18 @@
 import { getContent } from '../../open-authoring/github/api'
-import ContentNotFoundError from './ContentNotFoundError'
-import { b64DecodeUnicode } from '../../open-authoring/utils/base64'
 import OpenAuthoringError from '../../open-authoring/OpenAuthoringError'
+const atob = require('atob')
+
+const b64DecodeUnicode = str => {
+  // Going backwards: from bytestream, to percent-encoding, to original string.
+  return decodeURIComponent(
+    atob(str)
+      .split('')
+      .map(function(c) {
+        return '%' + ('00' + c.charCodeAt(0).toString(16)).slice(-2)
+      })
+      .join('')
+  )
+}
 
 // TODO - this name kinda sucks,
 // Throw a formatted error on 404, and decode github data properly
@@ -15,7 +26,7 @@ const getDecodedData = async (repoFullName, headBranch, path, accessToken) => {
 
   const errorStatus = response?.response?.status || 200
   if (errorStatus < 200 || errorStatus > 299) {
-    throw new OpenAuthoringError("Failed to get data.", errorStatus)
+    throw new OpenAuthoringError('Failed to get data.', errorStatus)
   }
 
   return { ...data, content: b64DecodeUnicode(data.content) }
