+++ utils/plugins/github-api/base64.ts
@@ -1,17 +1,4 @@
-const atob = require('atob')
 const btoa = require('btoa')
-
-export const b64DecodeUnicode = str => {
-  // Going backwards: from bytestream, to percent-encoding, to original string.
-  return decodeURIComponent(
-    atob(str)
-      .split('')
-      .map(function(c) {
-        return '%' + ('00' + c.charCodeAt(0).toString(16)).slice(-2)
-      })
-      .join('')
-  )
-}
 
 export const b64EncodeUnicode = str => {
   // first we use encodeURIComponent to get percent-encoded UTF-8,
@@ -19,7 +6,7 @@ export const b64EncodeUnicode = str => {
   // can be fed into btoa.
   return btoa(
     encodeURIComponent(str).replace(/%([0-9A-F]{2})/g, function toSolidBytes(
-      match,
+      _match,
       p1
     ) {
       return String.fromCharCode(parseInt(p1, 16))
